+++ src/main/java/com/fcgl/Listing/MessageQueueReceiveListing/State.java
@@ -1,0 +1,5 @@
+package com.fcgl.Listing.MessageQueueReceiveListing;
+
+public enum State {
+    INSERT, UPDATE, ERROR
+}
